backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
fake-useragent==1.4.0
tenacity==8.2.3
openai==1.3.5
langchain==0.0.350
tqdm==4.66.1
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
